+++ packages/neuralib-imaging/src/neuralib/widefield/__init__.py
@@ -1,3 +1,2 @@
 from .fft import *
-from .plot import *
 from .svd import *